build: use uint16_t instead of uint32_t for uint16_t attributes

+++ src/build.c
@@ -51,7 +51,7 @@ __build_u8(const struct nf_conntrack *ct, int a, struct nethdr *n, int b)
 static inline void 
 __build_u16(const struct nf_conntrack *ct, int a, struct nethdr *n, int b)
 {
-	uint32_t data = nfct_get_attr_u16(ct, a);
+	uint16_t data = nfct_get_attr_u16(ct, a);
 	data = htons(data);
 	addattr(n, b, &data, sizeof(uint16_t));
 }
